requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
requests==2.32.3
urllib3==2.2.3
Werkzeug==3.0.4
pdfplumber==0.11.4
pymongo==4.10.1
pandas==2.2.3
python-dotenv==1.0.1
